requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-docx>=0.8.11
streamlit>=1.28.0
pydantic>=2.0.0
PyYAML>=6.0
pytest>=7.0.0
